Remove redundant username parameter from API endpoints

+++ apps/api/src/controller/portfolioController.ts
@@ -1,3 +1,4 @@
+import * as mongoose from 'mongoose';
 import { isDocument } from '@typegoose/typegoose';
 import { PortfolioItemModel, TextItemModel } from '../models/portfolioItem';
 import { UserModel } from '../models/user';
@@ -27,12 +28,12 @@ const extractItemFromBody = (body?: PortfolioItemUnion): { model; item } => {
 interface Req {
   params: { username?: string; portfolioItemId?: string };
   body?: PortfolioItemUnion;
+  user?: { sub: string };
 }
 
 const createItem = async (req: Req, res: Res<never>) => {
-  const { username } = req.params;
   const { model, item } = extractItemFromBody(req.body);
-  if (model) {
+  if (model && req.user && req.user.sub) {
     const now = new Date();
     const newItem = await model.create({
       ...item,
@@ -41,7 +42,7 @@ const createItem = async (req: Req, res: Res<never>) => {
     });
     try {
       await UserModel.findOneAndUpdate(
-        { username },
+        { auth0Id: req.user.sub },
         { $push: { portfolio: newItem } }
       );
       res.sendStatus(201);
@@ -70,15 +71,23 @@ const viewItem = async (req: Req, res: Res<PortfolioItemUnion>) => {
 const editItem = async (req: Req, res: Res<never>) => {
   const { portfolioItemId } = req.params;
   const { model, item } = extractItemFromBody(req.body);
-  if (model) {
-    try {
-      await model.findByIdAndUpdate(portfolioItemId, {
-        ...item,
-        lastModified: new Date(),
-      });
-      res.sendStatus(200);
-    } catch {
-      res.sendStatus(404);
+  if (model && req.user && req.user.sub) {
+    const user = await UserModel.findOne({
+      auth0Id: req.user.sub,
+      portfolio: mongoose.Types.ObjectId(portfolioItemId),
+    });
+    if (user) {
+      try {
+        await model.findByIdAndUpdate(portfolioItemId, {
+          ...item,
+          lastModified: new Date(),
+        });
+        res.sendStatus(200);
+      } catch {
+        res.sendStatus(404);
+      }
+    } else {
+      res.sendStatus(401);
     }
   } else {
     res.sendStatus(400);
@@ -97,14 +106,22 @@ const viewAllItems = async (req: Req, res: Res<PortfolioItemUnion[]>) => {
 
 const deleteItem = async (req: Req, res: Res<never>) => {
   const { portfolioItemId } = req.params;
-  try {
-    await PortfolioItemModel.findByIdAndDelete(portfolioItemId);
-    /* Note: the deleted item ID will remain in the user's portfolio array,
-       but will not be returned from queries like `viewAllItems` as it is
-       filtered by `isDocument`. */
-    res.sendStatus(200);
-  } catch {
-    res.sendStatus(404);
+  const user = await UserModel.findOne({
+    auth0Id: req.user.sub,
+    portfolio: mongoose.Types.ObjectId(portfolioItemId),
+  });
+  if (user) {
+    try {
+      await PortfolioItemModel.findByIdAndDelete(portfolioItemId);
+      /* Note: the deleted item ID will remain in the user's portfolio array,
+         but will not be returned from queries like `viewAllItems` as it is
+         filtered by `isDocument`. */
+      res.sendStatus(200);
+    } catch {
+      res.sendStatus(404);
+    }
+  } else {
+    res.sendStatus(401);
   }
 };
